mlflow/requirements.txt
mlflow
pandas
scikit-learn
kafka-python
